+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Replacement `tsi1` indexes will be automatically generated on startup for shards
 1. [20535](https://github.com/influxdata/influxdb/pull/20535): Improve error message when opening BoltDB with unsupported file system options.
 1. [20542](https://github.com/influxdata/influxdb/pull/20542): Prevent extra output row from GROUP BY crossing DST boundary.
 1. [20615](https://github.com/influxdata/influxdb/pull/20615): Update Flux functions list in UI to reflect that `v1` package was renamed to `schema`.
+1. [20558](https://github.com/influxdata/influxdb/pull/20558): Prevent panic in `influxd upgrade` when V1 users exist and no V1 config is given.
 
 ## v2.0.3 [2020-12-14]
 ----------------------